uint32 is more efficient than uint8 on stm32

+++ USBMultiSerial.h
@@ -13,7 +13,7 @@ class USBSerialPort : public Stream {
 public:
     uint32 txPacketSize = USB_MULTI_SERIAL_DEFAULT_BUFFER_SIZE;
     uint32 rxPacketSize = USB_MULTI_SERIAL_DEFAULT_BUFFER_SIZE;
-    uint8 port;
+    uint32 port;
     virtual int available(void);// Changed to virtual
 
     uint32 read(uint8 * buf, uint32 len);
@@ -51,7 +51,7 @@ public:
     }    
 };
 
-template<const uint8 numPorts=3,const uint32 bufferSize=USB_MULTI_SERIAL_DEFAULT_BUFFER_SIZE>class USBMultiSerial {
+template<const uint32 numPorts=3,const uint32 bufferSize=USB_MULTI_SERIAL_DEFAULT_BUFFER_SIZE>class USBMultiSerial {
 private:
 	bool enabled = false;
     uint8 buffers[numPorts*2*bufferSize];
@@ -96,7 +96,7 @@ public:
     USBSerialPort ports[numPorts];
 
     USBMultiSerial() {
-        for (uint8 i=0;i<numPorts;i++) ports[i].setPort(i);
+        for (uint32 i=0;i<numPorts;i++) ports[i].setPort(i);
     }
 };
